NPM install should handle sudo/root constraints
When npm global installations are required to be run as sudo/root, the postinstall.sh file is denied execute permission by default by npm. To allow the script to execute we must set the --unsafe-perm config option.

This change also updates the cli tar.gz version to the most recent 1.13.7.