Add a simple example of using reason objects
Maybe I am just slow but when I first looked at the docs on objects I was stumped on how to create a simple object. After fooling around with it for about 20 min I finally figured out how they work. Coming from a JS background, objects don't have the idea of public and private methods in an explicit sense. In constructors, yes, but in plain ole' objects no, which is what I think of when I think of objects. I added a simple example, and referenced the other as advanced. I think the example I added shows how to create an object in its most simplest form while also referencing the very first example at the start of the doc. I think this would help newcomers, and advanced developers, to recognize that an object behaves much more like a constructor than an object and reduce confusion going forward in the doc.