Replace confusing ReasonReact instructions (#402)
If people try to follow the instructions on the page, and then enter the command shown to use the 'react' theme, they will encounter an error due to the fact that npm install hasn't been run. Additionally the instructions on this page, which describe running the code under node, really make no sense for ReasonReact. We'd really be better off just having them go to the ReasonReact instructions instead.