module.md: Clarify extending module signatures
The extending module signatures example does not compile unless a
concrete module `BaseComponent` is defined. Although these modules were
defined earlier in this document, they are now a bit separated and this
has caused some confusion for users trying to compile just this part.

This modifies the example to note that a normal `include` will work for
extending another module type and provides a separate example of using
`include (module type of ModuleName)` to extend a module that doesn't
have a defined module type.