Recommend to use reason-language-server for emacs
I don't know exactly what was meant by upgrading reason-mode to use reason-language-server so I removed that text.
I'm pretty sure this wording is better than the status quo: even after hacking elisp and experimenting with merlin on the command line, I wasn't able to get it to recognize that my file was Reason code rather than OCaml.  With LSP it seems to be working fine for me so far, using the spacemacs layer (aside from merlin still running and generating errors sometimes -- I'll try and find time to investigate how to kill that off and update the spacemacs layer later on).